page transition moved to the main content
