+++ app/src/app/(pages)/addons/layout.tsx
@@ -1,21 +1,16 @@
+// app/(pages)/addons/layout.tsx
 import { ReactNode } from 'react';
 import Sidebar from '@/components/Sidebar';
 import { getAddonLinks } from '@/scripts/getAddonLinks';
+import MotionWrapper from '@/components/MotionWrapper'; // <- client wrapper
 
-interface AddonsLayoutProps {
-  children: ReactNode;
-}
+export default function AddonsLayout({ children }: { children: ReactNode }) {
+  const links = getAddonLinks(); // server-side
 
-const links = getAddonLinks();
-
-// Server component
-export default async function AddonsLayout({ children }: AddonsLayoutProps) {
   return (
-    <div className="flex w-full h-full">
+    <div className="flex w-full h-full fixed">
       <Sidebar links={links} type="addons" />
-
-      {/* Main content */}
-      <main className="flex-1 overflow-y-auto px-36 pt-16">{children}</main>
+      <MotionWrapper>{children}</MotionWrapper>
     </div>
   );
 }
