+++ app/src/app/(pages)/layout.tsx
@@ -1,21 +1,5 @@
-'use client';
+import { ReactNode } from 'react';
 
-import { motion } from 'motion/react';
-import { usePathname } from 'next/navigation';
-
-export default function PageLayout({ children }: { children: React.ReactNode }) {
-  const pathname = usePathname();
-
-  return (
-    <motion.div
-      key={pathname}
-      initial={{ opacity: 0, y: 10 }}
-      animate={{ opacity: 1, y: 0 }}
-      exit={{ opacity: 0, y: -10 }}
-      transition={{ duration: 0.3, ease: 'easeInOut' }}
-      className="boundary overflow-x-hidden"
-    >
-      {children}
-    </motion.div>
-  );
+export default function PagesLayout({ children }: { children: ReactNode }) {
+  return <main className="boundary bg-[rgb(var(--footer-bg))]">{children}</main>;
 }
